pyproject.toml
[project]
name = "mootboard"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "google>=3.0.0",
    "google-adk>=1.18.0",
    "google-genai>=1.47.0",
    "google-generativeai>=0.8.5",
    "pillow>=11.3.0",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.20",
]
